Add appointment API router
